% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeDESeq2.R
\name{pipeDESeq2}
\alias{pipeDESeq2}
\title{A pipeline for DESeq2}
\usage{
pipeDESeq2(counts, info, formula = NULL, reduced = NULL, testNames = NULL,
  contrasts = NULL, contrastFormula = NULL, contrastNames = NULL,
  returnNormData = "none", geneIDs = NA, verbose = TRUE,
  fitType = "parametric", quiet = TRUE, betaPrior = F,
  modelMatrixType = "standard", parallel = F, minReplicatesForReplace = 7,
  ...)
}
\arguments{
\item{counts}{A count matrix, row names should be gene IDs}

\item{info}{An experimental design matrix}

\item{formula}{If specified, perform a likelihood ratio test. A vector of
character strings that can be coerced into a formula. Must be of the
same length as reduced. Specifies base formulas for LRT.}

\item{reduced}{A character vector of the same length as formula, which
can be coerced to a formula that represents a sub model to formula.
All reduced formulae must be sub models of the respective
formula.}

\item{testNames}{A character string of names for the LRT, if specified,
must be the same length as formula and reduced.}

\item{contrasts}{A list of character vectors that can be input into
DESeq2::results contrasts argument.}

\item{contrastFormula}{The formula specifying the design matrix for the
contrasts. Keep in mind that this formula specifies how the model is fit
and can affect inference of the effects of contrasts.}

\item{contrastNames}{Names for the contrasts. Must be the same length
as contrasts}

\item{returnNormData}{If NULL (default), no transformed data is returned.
Other options are "rlog" and "vst" which return rlog and variance stabilized data
respecitvel.}

\item{geneIDs}{The names of genes. If NA, use row names from counts matrix}

\item{verbose}{Logical, return progress updates?}

\item{fitType}{Passed to DESeq. Default is "parametric"}

\item{quiet}{Should messages from DESeq2 be suppressed?}

\item{betaPrior}{Passed to DESeq. Default is F}

\item{modelMatrixType}{Passed to DESeq. Default is "standard"}

\item{parallel}{Passed to DESeq. Default is F}

\item{minReplicatesForReplace}{Passed to DESeq. Default is 7}

\item{...}{additional arguments to pass to DESeq2::results}
}
\value{
a list with 3 elements (if test is not run, elements are NULL):
\itemize{
 \item{1. }{"LRT_results": The results from the likelihood ratio tests}
 \item{2. }{"CONT_results": The results from the contrasts}
 \item{3. }{"normData": normalized data}
}
}
\description{
\code{pipeDESeq2} Run a pipeline of DESeq2 functions for differential
gene expression.
}
\details{
This function runs the following pipeline:
\itemize{
 \item{1. }{DESeq's pipeline using the specified test}
 \item{2. }{Extraction of results (from DESeq2::results)}
 \item{3. }{Renaming of columns and combining results across tests}
}
}
\examples{
\dontrun{
library(deTools)
data(counts)
data(info)
stats<-pipeDESeq2(counts=counts, info=info,
   formula = c(" ~ trt"),
   reduced = c(" ~ 1"),
   testNames = c("trt"))
stats<-pipeDESeq2(counts=counts, info=info,
   contrastFormula = " ~ trt + geno",
   contrasts = list(c("geno","HAL2","FIL2")),
   contrastNames = "HvF")

stats<-pipeDESeq2(counts=counts, info=info,)
stats<-pipeDESeq2(counts=counts, info=info,
   formula = c(" ~ trt + geno + trt:geno"," ~ trt + geno", " ~ trt + geno"),
   reduced = c(" ~ trt + geno"," ~ trt"," ~ geno"),
   testNames = c("GxE","geno","trt"),
   contrastFormula = " ~ trt + geno",
   contrasts = list(c("trt","Wet","Dry"),c("geno","HAL2","FIL2")),
   contrastNames = c("WvD","HvF"),returnNormData = "vst")
}

}
